First implementation based on the cdflow module
Trying to keep generic.

We still hardcode the env variables for logspout and statsd.
We add an additional variable IMAGE only to be sure that the array of
variables is at least 1, so we will always return a json list of
elements.

For the labels it is not required, as the string in the json template
can be empty.

The object will return a unique json element, without the `[]`.