Allow define one optional volume
For some cases and workloads we need to pass some mountpoints to the container
definition. The mountpoints are passed as a list of maps in the json under
`mountPoints`[1].

In the related modules tf_task_definition, we added logic to define one
optional volume[2], and here we allow to specify the mounpoint for that
volume.

The mountpoint is defined as a map with the keys `sourceVolume`,
`containerPath` and (optional) `readOnly` .

[1] http://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#container_definition_storage
[2] https://github.com/mergermarket/tf_ecs_task_definition/pull/3